requirements.txt
dotenv==0.9.9
fastapi==0.129.0
google-cloud-bigquery==3.40.1
matplotlib==3.10.8
mysql-connector-python==9.6.0
numpy==2.4.2
packaging==26.0
pandas==2.3.3
pandas-gbq==0.33.0
plotly==6.5.2
plotly-express==0.4.1
pyarrow==23.0.1
python-dotenv==1.2.1
requests==2.32.5
SQLAlchemy==2.0.46
streamlit==1.54.0
uvicorn==0.41.0